SciResearcher/requirements.txt
# SciResearcher Requirements
# 基于 smolagents + Qwen3 + MinerU

# ========== 核心框架 ==========
smolagents>=1.0.0
litellm>=1.35.0

# ========== PDF解析 (MinerU) ==========
magic-pdf[full]==0.7.0b1

# ========== 向量数据库 ==========
chromadb>=0.4.0

# ========== API 客户端 ==========
requests>=2.31.0
openai>=1.0.0

# ========== 基础依赖 ==========
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# ========== 开发工具 (可选) ==========
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
